cloveros_settings.sh 4): check for internet

+++ home/user/cloveros_settings.sh
@@ -78,15 +78,19 @@ case "$choice" in
 		mkdir $tempdir
 		cd $tempdir
 		wget https://cloveros.ga/s/kernel.tar.xz
-		wget https://cloveros.ga/s/signatures/s/kernel.tar.xz.asc
-		sudo gpg --verify kernel.tar.xz.asc kernel.tar.xz
-		tar xf kernel.tar.xz
-		sudo mv initramfs-genkernel-*-gentoo* kernel-genkernel-*-gentoo* System.map-genkernel-*-gentoo* /boot/
-		sudo cp -R *-gentoo*/ /lib/modules/
+		if [[ -f kernel.tar.xz ]]; then
+			wget https://cloveros.ga/s/signatures/s/kernel.tar.xz.asc
+			sudo gpg --verify kernel.tar.xz.asc kernel.tar.xz
+			tar xf kernel.tar.xz
+			sudo mv initramfs-genkernel-*-gentoo* kernel-genkernel-*-gentoo* System.map-genkernel-*-gentoo* /boot/
+			sudo cp -R *-gentoo*/ /lib/modules/
+			sudo grub-mkconfig -o /boot/grub/grub.cfg
+			echo -e "\nKernel upgraded. (/boot/, /lib/modules/)"
+		else
+			echo -e "\nCould not retrieve file. Please connect to the Internet or try again."
+		fi
 		cd ..
 		rm -R $tempdir
-		sudo grub-mkconfig -o /boot/grub/grub.cfg
-		echo -e "\nKernel upgraded. (/boot/, /lib/modules/)"
 		;;
 
 	5)
